Version 4.426
git-svn-id: https://vtr-verilog-to-routing.googlecode.com/svn/trunk@2056 8e3573b8-cf2c-4f14-ef6d-137439e28b8b

+++ toro/TVPR_Interface/TVPR_IndexCount.cxx
@@ -200,7 +200,7 @@ void TVPR_IndexCount_c::ExtractString(
          if( this->siblingCount_ != SIZE_MAX )
          {
             char szSiblingCount[TIO_FORMAT_STRING_LEN_VALUE];
-            sprintf( szSiblingCount, "%lu", this->siblingCount_ );
+            sprintf( szSiblingCount, "%lu", static_cast< unsigned long >( this->siblingCount_ ));
 
             *psrIndexCount += " ";
             *psrIndexCount += "[";
